Generates n-grams from queries

Generates n-grams from queries

Usage

kwr_ngrams(
  x,
  max_words = 4,
  min_words = 1,
  min_n = 1,
  min_volume = 0,
  remove_nested = TRUE
)

Arguments

x

A kwresearch object, which queries will be n-grams calculated from, or a data frame of queries and volume.

max_words

Maximum number of words in n-grams.

min_words

Minimum number of words in n-grams.

min_n

Minimum number of queries. Only the n-grams with at least this number of queries will be included.

min_volume

Minimum search volume per n-gram. Only the n-grams with at least this volume will be included.

remove_nested

If TRUE, n-grams fully contained in another n-gram (number of queries must be the same) are filtered out from the result. May be slower.

Value

A tibble of n-grams with a basic stats (nuber of queries and sum of search volumes). The n-grams are orderd descendingly by number of queries and search volume. Use dplyr::arrange to change order. If stop words are set with the kwr_use_stopwords, they are removed from unigrams.

Examples

queries <- data.frame(
  query = c("seo", "keyword research"),
  volume = c(1000, 500)
)
kwr <- kwresearch(queries)
kwr |> kwr_ngrams()
#> # A tibble: 2 × 3
#>   token                n volume
#>   <chr>            <int>  <dbl>
#> 1 seo                  1   1000
#> 2 keyword research     1    500